ENH: skeleton module for testing injections, and some refactorings in top imports
