+++ duecredit/tests/mod/__init__.py
@@ -1,0 +1,4 @@
+"""Module to test various duecredit functionality, e.g. injections"""
+
+from .imported import *
+
