+++ duecredit/tests/mod/imported.py
@@ -1,0 +1,4 @@
+def testfunc1(arg1, kwarg1=None):
+    return "testfunc1: %s, %s" % (arg1, kwarg1)
+
+
